Remove static ConverterFactory.
